+++ company-site/app/components/HeadLine.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { ReactElement, ReactText } from 'react'
 
 export default function HeadLine({
