+++ company-site/app/components/contents/home.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Link from 'next/link'
 import { ReactElement } from 'react'
 import HeadLine from '../HeadLine'
