+++ company-site/app/components/contents/house.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { ReactElement, useState } from 'react'
 import HeadLine from '../HeadLine'
 import {
