+++ company-site/app/components/contents/surroundings.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { ReactElement } from 'react'
 import HeadLine from '../HeadLine'
