+++ company-site/app/components/floorPlan/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { ReactElement } from 'react'
 import { Room, RoomNames, Rooms } from '../../domains/house'
