+++ company-site/app/components/icons/menu.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { ReactElement } from 'react'
 
 export default function MenuIcon(): ReactElement {
