+++ company-site/app/page.tsx
@@ -1,0 +1,8 @@
+'use client'
+
+import { ReactElement } from 'react'
+import HomePage from './HomePage'
+
+export default function Page(): ReactElement {
+  return HomePage()
+}
